+++ compiler/core/lam.ml
@@ -581,8 +581,7 @@ let prim ~primitive:(prim : Lam_primitive.t) ~args loc : t =
         | ( f :: fields,
             Lprim
               {
-                primitive =
-                  Pfield (pos, Fld_module {name = f1; jsx_component = false});
+                primitive = Pfield (pos, Fld_module {name = f1});
                 args = [(Lglobal_module (v1, _) | Lvar v1)];
               }
             :: args ) ->
@@ -593,8 +592,7 @@ let prim ~primitive:(prim : Lam_primitive.t) ~args loc : t =
       | ( field1 :: rest,
           Lprim
             {
-              primitive =
-                Pfield (pos, Fld_module {name = f1; jsx_component = false});
+              primitive = Pfield (pos, Fld_module {name = f1});
               args = [((Lglobal_module (v1, _) | Lvar v1) as lam)];
             }
           :: args1 ) ->