+++ compiler/ml/translmod.ml
@@ -64,20 +64,14 @@ let rec apply_coercion loc strict (restr : Typedtree.module_coercion) arg =
   | Tcoerce_structure (pos_cc_list, id_pos_list, runtime_fields) ->
     Lambda.name_lambda strict arg (fun id ->
         let get_field_name name pos =
-          Lambda.Lprim
-            ( Pfield (pos, Fld_module {name; jsx_component = false}),
-              [Lvar id],
-              loc )
+          Lambda.Lprim (Pfield (pos, Fld_module {name}), [Lvar id], loc)
         in
         let lam =
           Lambda.Lprim
             ( Pmakeblock (Blk_module runtime_fields),
               Ext_list.map2 pos_cc_list runtime_fields (fun (pos, cc) name ->
                   apply_coercion loc Alias cc
-                    (Lprim
-                       ( Pfield (pos, Fld_module {name; jsx_component = false}),
-                         [Lvar id],
-                         loc ))),
+                    (Lprim (Pfield (pos, Fld_module {name}), [Lvar id], loc))),
               loc )
         in
         wrap_id_pos_list loc id_pos_list get_field_name lam)
@@ -438,10 +432,7 @@ and transl_structure loc fields cc rootpath final_env = function
                 Pgenval,
                 id,
                 Lprim
-                  ( Pfield
-                      ( pos,
-                        Fld_module {name = Ident.name id; jsx_component = false}
-                      ),
+                  ( Pfield (pos, Fld_module {name = Ident.name id}),
                     [Lvar mid],
                     incl.incl_loc ),
                 body ),