+++ tests/build_tests/react_ppx/src/abstract_component_test.res.js
@@ -25,6 +25,8 @@ let GenericRenderProp = {
 };
 
 export {
+  PolyVariantLowerBound,
+  GenericRenderProp,
   Abstract_component_test$PolyVariantLowerBound,
   Abstract_component_test$GenericRenderProp,
 }